# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank system
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags:
      - Authentication
      operationId: updateAccessToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: Successfully refreshed token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtTokenDto"
        "400":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Invalid refresh token
                code: 400
        "401":
          description: Refresh token expired or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Refresh token expired or invalid
                code: 401
  /auth/login:
    post:
      tags:
      - Authentication
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequestDto"
        required: true
      responses:
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtTokenDto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Validation failed
                code: 400
                validationErrors:
                - field: username
                  message: Username must not be empty
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Invalid username or password
                code: 401
  /admin/cards:
    get:
      tags:
      - Admin Card Management
      summary: Get card
      description: Get all cards
      operationId: getAllCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "validityPeriod,ASC"
          items:
            type: string
      responses:
        "200":
          description: Card found
          content:
            application/json: {}
    post:
      tags:
      - Admin Card Management
      summary: Create a new card for a user
      description: Creates a new bank card for a specified user
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateDto"
        required: true
      responses:
        "201":
          description: Card created successfully
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Validation failed
                code: 400
                validationErrors:
                - field: cardNumber
                  message: Incorrect card number format
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "User by id: 1 not found"
                code: 404
        "409":
          description: Card already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by number: 4000123456789010 already exists"
                code: 409
...
  /admin/cards/{cardId}/status:
    patch:
      tags:
      - Admin Card Management
      summary: Update card status
      description: "Updates the status of a card (e.g., ACTIVE, BLOCKED)"
      operationId: updateStatusCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusCardDto"
        required: true
      responses:
        "200":
          description: Card status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
...
  /cards:
    get:
      tags:
      - User Card Management
      summary: Get all user cards
      description: Retrieves a paginated list of all cards belonging to the authenticated
        user
      operationId: getUserCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "balance,ASC"
          items:
            type: string
      responses:
        "200":
          description: Cards found
          content:
            application/json: {}
  /cards/{id}:
    get:
      tags:
      - User Card Management
      summary: Get card by ID
      description: "Retrieves a specific card by its ID, if it belongs to the authenticated\
        \ user"
      operationId: getCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found for user 1"
                code: 404
...
  /admin/cards/{id}:
    get:
      tags:
      - Admin Card Management
      summary: Get card by ID
      description: Retrieves card details by its ID
      operationId: getCardById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
    delete:
      tags:
      - Admin Card Management
      summary: Delete card by ID
      description: Deletes a card by its ID
      operationId: deleteCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Card deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
  /admin/cards/user/{id}:
    get:
      tags:
      - Admin Card Management
      summary: Get all cards for a user
      description: Retrieves a paginated list of all cards for a specified user
      operationId: getUserCards_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "validityPeriod,ASC"
          items:
            type: string
      responses:
        "200":
          description: Cards found
          content:
            application/json: {}
  /admin/cards/by-username:
    get:
      tags:
      - Admin Card Management
      summary: Get card by username
      description: Get all the user's cards by name
      operationId: getCardsByUserName
      parameters:
      - name: username
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "validityPeriod,ASC"
          items:
            type: string
      responses:
        "200":
          description: Card found
          content:
            application/json: {}
  /admin/cards/by-number:
    get:
      tags:
      - Admin Card Management
      summary: Get card by card number
      description: Retrieves card details by its number
      operationId: getCardByCardNumber
      parameters:
      - name: number
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by: **** **** **** 9010 not found"
                code: 404
    delete:
      tags:
      - Admin Card Management
      summary: Delete card by card number
      description: Deletes a card by its number
      operationId: deleteCardByCardNumber
      parameters:
      - name: number
        in: query
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Card deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by number: 4000123456789010 not found"
                code: 404
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
        message:
          type: string
          description: Error message
          example: Validation failed
        code:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        validationErrors:
          type: array
          description: List of validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      description: Details of a validation error
      properties:
        field:
          type: string
          description: Field where the error occurred
          example: username
        message:
          type: string
          description: Error message for the field
          example: Username must not be empty
...
    CardDto:
      type: object
      description: Card details
      properties:
        cardId:
          type: integer
          format: int64
          description: Card ID
          example: 1
        cardNumber:
          type: string
          description: Masked card number
          example: '**** **** **** 9010'
        validityPeriod:
          type: string
          format: date
          description: Expiration date of the card
          example: 2028-12-31
        statusCard:
          type: string
          description: Current status of the card
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - PENDING_BLOCKED
        balance:
          type: number
          description: Current balance of the card
          example: 1000.0
...
    RefreshTokenRequest:
      type: object
      description: Request to refresh an access token
      properties:
        refreshToken:
          type: string
          description: Refresh token
          example: def50200f496c85c9a1c...
          minLength: 1
      required:
      - refreshToken
    JwtTokenDto:
      type: object
      description: JWT token response
      properties:
        accessToken:
          type: string
          description: Access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refreshToken:
          type: string
          description: Refresh token
          example: def50200f496c85c9a1c...
    AuthRequestDto:
      type: object
      description: Request body for user authentication
      properties:
        username:
          type: string
          description: User's login name
          example: john_doe
          minLength: 1
        password:
          type: string
          description: User's password
          example: password123
          minLength: 1
      required:
      - password
      - username
    CardCreateDto:
      type: object
      description: Data for creating a new card
      properties:
        userId:
          type: integer
          format: int64
          description: User ID of the card holder
          example: 1
        cardNumber:
          type: string
          description: Card number
          example: 4000123456789010
        validityPeriod:
          type: string
          format: date
          description: Expiration date of the card
          example: 2028-12-31
      required:
      - cardNumber
      - userId
      - validityPeriod
...
    UpdateStatusCardDto:
      type: object
      description: Request to update a card status
      properties:
        newStatus:
          type: string
          description: New status for the card
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - PENDING_BLOCKED
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT